Additional Base_Rowset functionality
Simplified some code that was converting column name to method names and
then calling the methods. This can all just be done by going through
__set or __get.

Added getRowByPrimaryKeys to rowset so you can request a specific row
from a rowset. Just does the looping and searching for you. Makes use of
filterBy to do the work.

Added save to rowset which loops through all children and calls save.

Added delete to rowset which loops through all children and deletes.

setFromArray allows you to pass in an array of arrays to set data on
your rows. If matches by primary key if it can by calling
getRowByPrimaryKeys. If not, it creates a new row and adds it to the
rowset.

__call routes unknown get or set methods to the individual rows to
return an array of values or set all values for each rowset.

Base_Table::createRowset was modified so you can pass $data to it which
then calls $rowset->setFromArray($data);

Base_Row::setFromArray was updated to attempt to call any other setting
methods for data that isn't part of the original table but was still
passed to setFromArray.