Fixes ip address converters when loading from database
Changes setIpAddress logic to not convert ip addresses that may have
already be converted from ip2long.

Updates _refresh of row method to call toInternalArray() instead of
toArray() so we still get th raw database values loaded in the object
when we _refresh. This was a problem because toArray() had been updated
to go through the getters and setters that had their own custom logic as
seen with ip address comment fields.