Changes _refresh to not requery the database
In master/slave database setups, calling a read write after doing
an insert was throwing an exception about the parent object not
being found. This was a result of trying to read the newly inserted
record from a slave database before it got there.

This changes _refresh to not read from the database to confirm the
write has happened. Instead, we copy $_data down to $_cleanData
and clear the $_modifiedFields properity.