adds pagination to scaffold table and rowset
Brings in a series of methods to access pagination data on the rowset
and a method in the table to apply pagination to a select statement.
Modifies code generation so findBy methods receive optional options parameter
and pass it to findByColumnValues.
Modifies findByColumnValue(s) to receive optional options parameter and call
paginate() instead of fetchAll() if options contains 'pagination' key.
Adds findAll() method that utilizes new pagination functionality.